responding with 404 on Not Found and with 400 (bad request) when actor is in use.

+++ app_movies_orm.py
@@ -1,7 +1,7 @@
 # ORM version
 from typing import Any
 
-from fastapi import FastAPI
+from fastapi import FastAPI, HTTPException
 from playhouse.shortcuts import model_to_dict
 
 from orm_models import Actor, Movie
@@ -17,14 +17,14 @@ async def get_orm_movies():
 async def get_orm_single_movie(movie_id:int):
     movie = Movie.get_or_none(Movie.id == movie_id)
     if movie is None:
-        return {"message": "Movie not found"}
+        raise HTTPException(status_code=404, detail="Movie not found")
     return model_to_dict(movie)
 
 @app_movies_orm.get("/movies/{movie_id}/actors")
 async def get_actors_for_movie(movie_id:int):
     movie = Movie.get_or_none(Movie.id == movie_id)
     if movie is None:
-        return {"message": "Movie not found"}
+        raise HTTPException(status_code=404, detail="Movie not found")
     return [model_to_dict(actor) for actor in movie.actors]
 
 @app_movies_orm.post("/movies")
@@ -37,7 +37,7 @@ def delete_orm_movie(movie_id: int):
     query = Movie.delete().where(Movie.id == movie_id)
     count = query.execute()
     if count == 0:
-        return {"message": f"no movie found by id: {movie_id}"}
+        raise HTTPException(status_code=404, detail=f"no movie found by id: {movie_id}")
     return {"message": f"Movie with ID {movie_id} deleted successfully."}
 
 @app_movies_orm.delete("/movies")
@@ -51,7 +51,7 @@ def update_orm_movie(movie_id: int, params: dict[str, Any]):
     query = Movie.update(**params).where(Movie.id == movie_id)
     updated = query.execute()
     if not updated:
-        return {"message": "Movie not found"}
+        raise HTTPException(status_code=404, detail="Movie not found")
     return {"message": f"Movie with ID {movie_id} updated successfully."}
 
 @app_movies_orm.get("/actors")
@@ -63,7 +63,7 @@ async def get_orm_actors():
 async def get_orm_single_actor(actor_id:int):
     actor = Actor.get_or_none(Actor.id == actor_id)
     if actor is None:
-        return {"message": "Actor not found"}
+        raise HTTPException(status_code=404, detail="Actor not found")
     return model_to_dict(actor)
 
 @app_movies_orm.post("/actors")
@@ -75,7 +75,7 @@ async def add_orm_actor(params: dict[str, Any]):
 async def update_orm_actor(id: int, params: dict[str, Any]):
     actor = Actor.get_or_none(Actor.id == id)
     if actor is None:
-        return {"message": "Actor not found"}
+        raise HTTPException(status_code=404, detail="Actor not found")
     for key, value in params.items():
         setattr(actor, key, value)
     actor.save()
@@ -85,7 +85,10 @@ async def update_orm_actor(id: int, params: dict[str, Any]):
 async def delete_orm_actor(id: int):
     actor = Actor.get_or_none(Actor.id == id)
     if actor is None:
-        return {"message": "Actor not found"}
-    # TODO: block deletion if actor is still in use
+        raise HTTPException(status_code=404, detail="Actor not found")
+    
+    if actor.movies.count() > 0:
+        raise HTTPException(status_code=400, detail=f"Actor {actor.name} {actor.surname} can not be removed, as it is connected to movies.")
+
     actor.delete_instance()
     return {"message": f"Actor with ID {id} deleted successfully."}
